feat/create CatInfoItem, add adaptive to wrapper class of MyKitten component
