+++ src/components/Account/AccountBlocks/MyKitten/MyKitten.module.scss
@@ -4,170 +4,27 @@
 .wrapper {
 	width: 100%;
 	background-color: #fff;
+	padding-bottom: 5.88rem;
+
+	@include tablet {
+		gap: 7rem;
+	}
+
+	@include full_desktop {
+		gap: 16.56rem;
+	}
 }
 
 .kittensList {
 	display: flex;
 	flex-direction: column;
-}
-
-.kittenItem {
-	display: flex;
-	flex-direction: column;
-
-	@include full_desktop {
-		display: grid;
-		grid-template-columns: 26.375rem 1fr;
-		grid-auto-rows: auto;
-	}
-}
-
-.kittenTitle {
-	color: $color-primary-orange;
-	font-family: $secondary-font;
-	font-size: 1.375rem;
-	font-weight: 900;
-	line-height: 110%;
-	text-align: start;
-	width: 17.8125rem;
+	gap: 4.19rem;
 
 	@include tablet {
-		width: 26.375rem;
-		font-size: 2.5rem;
-	}
-}
-
-.kittenDescrBody {
-	display: flex;
-	flex-direction: column;
-	margin: 0.75rem 0;
-	padding: 0.75rem 0;
-
-	@include tablet {
-		margin: 1.5rem 0;
+		gap: 5.25rem;
 	}
 
 	@include full_desktop {
-		grid-column: 1 / 3;
-	}
-}
-
-.kittenId {
-	color: $color-primary-green;
-	font-family: $primary-font;
-	font-size: 1rem;
-	font-weight: 400;
-	line-height: 120%; /* 1.2rem */
-
-	@include tablet {
-		font-weight: 500;
-	}
-}
-
-.kittenAge {
-	margin-top: 0.75rem;
-	color: $color-primary-green;
-	font-family: $primary-font;
-	font-size: 1.25rem;
-	font-weight: 700;
-	line-height: 120%; /* 1.5rem */
-
-	@include tablet {
-		font-weight: 900;
-	}
-}
-
-.kittenBirthday {
-	color: $color-primary-green;
-	font-family: $primary-font;
-	font-size: 1rem;
-	font-weight: 400;
-	line-height: 120%; /* 1.2rem */
-
-	@include tablet {
-		font-weight: 900;
-	}
-}
-
-.timerBlock {
-	display: flex;
-	flex-direction: column;
-	gap: 0.75rem;
-
-	@include full_desktop {
-		grid-row: 1 / 2;
-		grid-column: 2 / 3;
-	}
-}
-
-.timerTitle {
-	color: $color-primary-dark-green;
-	text-align: center;
-	font-family: $primary-font;
-	font-size: 1rem;
-	font-weight: 700;
-	line-height: 132%; /* 1.32rem */
-
-	@include tablet {
-		font-size: 1.125rem;
-	}
-
-	@include full_desktop {
-		font-size: 1.125rem;
-	}
-}
-
-.timerWrapper {
-	display: flex;
-	justify-content: center;
-	gap: 1.19rem;
-}
-
-.timerUnit {
-	display: grid;
-	grid-template-columns: repeat(2, 1fr);
-	grid-template-rows: repeat(2, auto);
-	column-gap: 0.25rem;
-}
-
-.timerNum {
-	padding: 0.5rem 0.625rem;
-	background: $color-primary-green;
-	color: $color-system-default-0;
-	font-family: $tertiary-font;
-	font-size: 1.5rem;
-	font-weight: 600;
-	line-height: normal;
-}
-
-.timerNoun {
-	grid-column: 1 / 3;
-	text-align: center;
-	color: $color-primary-dark-green;
-	font-family: $primary-font;
-	font-size: 1.125rem;
-	font-weight: 400;
-	line-height: 150%; /* 1.6875rem */
-}
-
-.sliderBlock {
-	@include full_desktop {
-		grid-column: 1 / 3;
-	}
-}
-
-.btnWrapper {
-	margin: 0 auto;
-	margin-top: 3.25rem;
-	width: 100%;
-
-	@include small_tablet {
-		margin-top: 2.75rem;
-		width: auto;
-	}
-
-	@include full_desktop {
-		margin-top: 2.5rem;
-		grid-column: 1 / 3;
+		gap: 6.12rem;
 	}
 }
